+++ src/apify/storage_clients/__init__.py
@@ -2,11 +2,11 @@ from crawlee.storage_clients import MemoryStorageClient
 
 from ._apify import ApifyStorageClient
 from ._file_system import ApifyFileSystemStorageClient as FileSystemStorageClient
-from ._hybrid_apify import ApifyHybridStorageClient
+from ._smart_apify import SmartApifyStorageClient
 
 __all__ = [
-    'ApifyHybridStorageClient',
     'ApifyStorageClient',
     'FileSystemStorageClient',
     'MemoryStorageClient',
+    'SmartApifyStorageClient',
 ]
