+++ src/apify/storage_clients/_smart_apify/__init__.py
@@ -1,0 +1,1 @@
+from ._storage_client import SmartApifyStorageClient
