+++ tests/integration/test_apify_storages.py
@@ -7,7 +7,7 @@ from crawlee.storages import Dataset, KeyValueStore, RequestQueue
 
 from .conftest import MakeActorFunction, RunActorFunction
 from apify import Actor, Configuration
-from apify.storage_clients import ApifyHybridStorageClient, ApifyStorageClient, MemoryStorageClient
+from apify.storage_clients import ApifyStorageClient, MemoryStorageClient, SmartApifyStorageClient
 
 
 @pytest.mark.parametrize(
@@ -38,13 +38,13 @@ async def test_alias_concurrent_creation_local(
 async def test_actor_full_explicit_storage_init(apify_token: str) -> None:
     service_locator.set_configuration(Configuration(token=apify_token))
     service_locator.set_storage_client(
-        ApifyHybridStorageClient(
+        SmartApifyStorageClient(
             local_storage_client=MemoryStorageClient(),
             cloud_storage_client=ApifyStorageClient(request_queue_access='shared'),
         )
     )
-    async with Actor():
-        # If service locator was already set with ApifyHybridStorageClient, the actor will use it.
+    async with Actor:
+        # If service locator was already set with SmartApifyStorageClient, the actor will use it.
         # Storages should be different when force_cloud is used outside the Apify platform
         assert await Actor.open_dataset() is not await Actor.open_dataset(force_cloud=True)
         assert await Actor.open_key_value_store() is not await Actor.open_key_value_store(force_cloud=True)
@@ -54,13 +54,13 @@ async def test_actor_full_explicit_storage_init(apify_token: str) -> None:
 async def test_actor_full_explicit_storage_init_same_client(apify_token: str) -> None:
     service_locator.set_configuration(Configuration(token=apify_token))
     service_locator.set_storage_client(
-        ApifyHybridStorageClient(
+        SmartApifyStorageClient(
             local_storage_client=ApifyStorageClient(request_queue_access='shared'),
             cloud_storage_client=ApifyStorageClient(request_queue_access='shared'),
         )
     )
-    async with Actor():
-        # If service locator was already set with ApifyHybridStorageClient, the actor will use it.
+    async with Actor:
+        # If service locator was already set with SmartApifyStorageClient, the actor will use it.
         # Storages should be same as the equivalent storage client is for both local and cloud storage client
         assert await Actor.open_dataset() is await Actor.open_dataset(force_cloud=True)
         assert await Actor.open_key_value_store() is await Actor.open_key_value_store(force_cloud=True)
@@ -70,9 +70,9 @@ async def test_actor_full_explicit_storage_init_same_client(apify_token: str) ->
 async def test_actor_partial_explicit_cloud_storage_init(apify_token: str) -> None:
     service_locator.set_configuration(Configuration(token=apify_token))
     service_locator.set_storage_client(ApifyStorageClient(request_queue_access='shared'))
-    async with Actor():
+    async with Actor:
         # If service locator was already set with ApifyStorageClient, the actor will use it as cloud_storage_client of
-        # ApifyHybridStorageClient
+        # SmartApifyStorageClient
         assert await Actor.open_dataset() is not await Actor.open_dataset(force_cloud=True)
         assert await Actor.open_key_value_store() is not await Actor.open_key_value_store(force_cloud=True)
         assert await Actor.open_request_queue() is not await Actor.open_request_queue(force_cloud=True)
@@ -81,9 +81,9 @@ async def test_actor_partial_explicit_cloud_storage_init(apify_token: str) -> No
 async def test_actor_partial_explicit_local_storage_init(apify_token: str) -> None:
     service_locator.set_configuration(Configuration(token=apify_token))
     service_locator.set_storage_client(MemoryStorageClient())
-    async with Actor():
+    async with Actor:
         # If service locator was already set with non-ApifyStorageClient, the actor will use it as local_storage_client
-        # of ApifyHybridStorageClient
+        # of SmartApifyStorageClient
         assert await Actor.open_dataset() is not await Actor.open_dataset(force_cloud=True)
         assert await Actor.open_key_value_store() is not await Actor.open_key_value_store(force_cloud=True)
         assert await Actor.open_request_queue() is not await Actor.open_request_queue(force_cloud=True)
@@ -91,7 +91,7 @@ async def test_actor_partial_explicit_local_storage_init(apify_token: str) -> No
 
 async def test_actor_implicit_storage_init(apify_token: str) -> None:
     service_locator.set_configuration(Configuration(token=apify_token))
-    async with Actor():
+    async with Actor:
         assert await Actor.open_dataset() is not await Actor.open_dataset(force_cloud=True)
         assert await Actor.open_key_value_store() is not await Actor.open_key_value_store(force_cloud=True)
         assert await Actor.open_request_queue() is not await Actor.open_request_queue(force_cloud=True)
@@ -103,15 +103,15 @@ async def test_actor_full_explicit_storage_init_on_platform(
     async def main() -> None:
         from crawlee import service_locator
 
-        from apify.storage_clients import ApifyHybridStorageClient, ApifyStorageClient, MemoryStorageClient
+        from apify.storage_clients import ApifyStorageClient, MemoryStorageClient, SmartApifyStorageClient
 
         service_locator.set_storage_client(
-            ApifyHybridStorageClient(
+            SmartApifyStorageClient(
                 local_storage_client=MemoryStorageClient(),
                 cloud_storage_client=ApifyStorageClient(request_queue_access='shared'),
             )
         )
-        async with Actor():
+        async with Actor:
             # Storages should be same as the cloud client is used on the platform
             assert await Actor.open_dataset() is await Actor.open_dataset(force_cloud=True)
             assert await Actor.open_key_value_store() is await Actor.open_key_value_store(force_cloud=True)
