+++ tests/unit/actor/test_configuration.py
@@ -11,7 +11,7 @@ from crawlee.storage_clients import FileSystemStorageClient
 
 from apify import Actor
 from apify import Configuration as ApifyConfiguration
-from apify.storage_clients._hybrid_apify._storage_client import ApifyHybridStorageClient
+from apify.storage_clients._smart_apify._storage_client import SmartApifyStorageClient
 
 
 @pytest.mark.parametrize(
@@ -111,8 +111,8 @@ async def test_crawler_implicit_local_storage() -> None:
     async with Actor():
         crawler = BasicCrawler()
 
-    assert isinstance(service_locator.get_storage_client(), ApifyHybridStorageClient)
-    assert isinstance(crawler._service_locator.get_storage_client(), ApifyHybridStorageClient)
+    assert isinstance(service_locator.get_storage_client(), SmartApifyStorageClient)
+    assert isinstance(crawler._service_locator.get_storage_client(), SmartApifyStorageClient)
 
 
 async def test_crawlers_own_configuration(tmp_path: Path) -> None:
